fix: move client to Greencity folder
