+++ green_city_sh.Tests/green_city_sh.Tests.csproj
@@ -46,8 +46,4 @@
     </None>
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Api\Clients\Greencity\" />
-  </ItemGroup>
-
 </Project>
